make typescript happy
